interface: reorganized groups directory structure to have like-kind components together
