+++ pkg/interface/src/views/landscape/components/ResourceSkeleton.tsx
