+++ pkg/interface/src/views/landscape/components/Sidebar/Apps.tsx
