+++ pkg/interface/src/views/landscape/components/Sidebar/Sidebar.tsx
@@ -17,7 +17,7 @@ import { SidebarInvite } from "./SidebarInvite";
 import GlobalApi from "~/logic/api/global";
 import { AppName } from "~/types/noun";
 import { alphabeticalOrder } from "~/logic/lib/util";
-import { GroupSwitcher } from "~/views/apps/groups/components/GroupSwitcher";
+import { GroupSwitcher } from "../GroupSwitcher";
 import {
   AppInvites,
   Associations,
