+++ pkg/interface/src/views/landscape/components/Sidebar/SidebarList.tsx
